Allow To Parse Unquoted Email Local Part That Has Space
This commit allows space to be consumed as atom. In RFC spec,
space is a special character which can not appear in un-quoted
local-part of an email. However, there are emails failed to follow
this RFC, and this commit modification allow this kind of broken
email to be parsed out as well. The correction code is inspired by
Thunderbird parser where any suffix and prefix space of local-part
of such email will be removed and if there is still space within
this un-quoted local-part email, then it will be quoted.